+++ core/pioreactor/web/static/static/Check-B7om9gAH.js
@@ -1,0 +1,1 @@
+import{c as o,j as s}from"./index-AyobIhZO.js";const c=o(s.jsx("path",{d:"M9 16.17 4.83 12l-1.42 1.41L9 19 21 7l-1.41-1.41z"}));export{c as C};